Partial refactoring.  Still needs updates for new Spring MVC version
