+++ deliverytiersuite/delivery-tier-suite/p13n-ds/src/main/java/com/percussion/soln/p13n/tracking/ds/web/DummyController.java
@@ -4,6 +4,7 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.servlet.ModelAndView;
 
 @Controller
@@ -22,8 +23,9 @@ public class DummyController {
     /*
      * Wrapper for ProfileMail.jsp
      */
+    @RequestMapping
     public ModelAndView handleRequest(HttpServletRequest request,
-            HttpServletResponse response) throws Exception {
+            HttpServletResponse response) {
         return new ModelAndView(getView());
     }
