+++ deliverytiersuite/delivery-tier-suite/p13n-ds/src/test/java/com/percussion/soln/p13n/delivery/ds/DSDeliveryContextFactoryTest.java
@@ -1,4 +1,4 @@
-package test.percussion.soln.p13n.delivery.ds;
+package com.percussion.soln.p13n.delivery.ds;
 
 import static java.util.Arrays.*;
 import static org.junit.Assert.*;
