+++ deliverytiersuite/delivery-tier-suite/p13n-ds/src/test/java/com/percussion/soln/p13n/delivery/web/ds/DeliveryControllerTest.java
@@ -1,4 +1,4 @@
-package test.percussion.soln.p13n.delivery.web.ds;
+package com.percussion.soln.p13n.delivery.web.ds;
 
 import static java.util.Arrays.*;
 import static org.junit.Assert.*;
