+++ deliverytiersuite/delivery-tier-suite/p13n-ds/src/test/java/com/percussion/soln/p13n/tracking/VisitorProfileTest.java
@@ -1,4 +1,4 @@
-package test.percussion.soln.p13n.tracking;
+package com.percussion.soln.p13n.tracking;
 
 import static org.junit.Assert.*;
